Fighting to get entity_ids on screen for debugging.

+++ resources/shaders/forward_pass.glsl
@@ -25,7 +25,7 @@ void main() {
 // Output buffers (Textures)
 layout(location=0) out vec4 out_fragment_color;
 layout(location=1) out vec4 out_fragment_normal;
-layout(location=2) out int out_fragment_entity_id;
+layout(location=2) out float out_fragment_entity_id;
 
 // Input Buffers
 in vec3 v_normal;
@@ -73,7 +73,7 @@ void main() {
 
     out_fragment_color = vec4(c * 0.5, uColor.a);
     out_fragment_normal = vec4(n, 1.0);
-    out_fragment_entity_id = entity_id;
+    out_fragment_entity_id = float(entity_id);
 }
 
 #endif
